Implemented addStudent, returnStudents and sendEmailToStudent

+++ server/controllers/userController.js
@@ -3,16 +3,22 @@ import {createRequire} from 'module'
 const require = createRequire(import.meta.url)
 import express from "express"
 const passport = require("passport")
-const AWS = require('aws-sdk')
+import s3 from "../aws/s3.js"
 // const GoogleStrategy = require('passport-google-oauth20').Strategy;
 // const findOrCreate = require('mongoose-findorcreate');
 import User from "../models/userModel.js"
 import Otp from "../models/otpModel.js"
 const {Auth} = require("two-step-auth")
+import nodemailer from "nodemailer"
 
-const s3 = new AWS.S3({
-  apiVersion: '2006-03-01'
-})
+const transporter = nodemailer.createTransport({
+  service: 'gmail',
+  auth: {
+    user: 'readpalishere@gmail.com',
+    pass: 'youcannotbreakit'
+  }
+});
+
 const router = express.Router()
 
 const userLogin = (req, res, next) => {
@@ -77,8 +83,10 @@ const userSignup = async (req, res) => {
 }
 
 
-
 const verifyEmailForSignup = async (req, res) => {
+  if(!req.body.email) {
+    return res.json({message: "missing body"})
+  }
   const foundUser = await User.findOne({username: req.body.email.toLowerCase()}).catch(err => {console.log("Finding user error: " + err)})
   if (foundUser == null) {
     const result = await Auth(req.body.email.toLowerCase(), "ReadPal").catch(err => {console.log("Sending Otp error when verifying email for signup: " + err)})
@@ -112,6 +120,12 @@ const verifyEmailForSignup = async (req, res) => {
 
 const addProfile = async (req, res) => {
   const userID = req.params.userID
+
+
+  if(!req.body.age || !req.body.color || !req.body.animal) {
+    return res.json({message: "missing body"})
+  }
+
   const {
     age,
     color,
@@ -163,7 +177,71 @@ const addProfile = async (req, res) => {
 }
 
 
+const addStudent = async (req, res) => {
+  const userID = req.params.userID
+  try {
+    if(!req.body.email) {
+      return res.json({message: "missing body"})
+    }
+    const email = req.body.email.toLowerCase()
+
+    const foundUser = await User.findById({_id: userID}).catch(err => {console.log("Finding user error when adding student: " + err)})
+    const foundStudent = await User.findOne({username: email}).catch(err => {console.log("Finding student error when adding student: " + err)})
+
+    if(foundUser == null || foundStudent == null || foundUser.username == foundStudent.email) {
+      res.json({message: "invalid"})
+    }else {
+      const foundEmail = foundUser.students.find((student, index) => {
+        if(student.email == email) {
+          return true
+        }
+      })
+      if(foundEmail == undefined) {
+        foundUser.students.push({email: email})
+        const savedUser = await foundUser.save().catch(err => {console.log("Saving user error when adding student: " + err)})
+
+        const foundStudents = await User.findById(savedUser._id, 'students').catch(err => {console.log("Finding students error: " + err)})
+        const childObject = foundStudents.students.find((object, index) => { //search for student
+            if (object.email == email) {
+              return true
+            }
+          });
+          res.json({
+            message: "success",
+            studentID: childObject._id
+          })
+      }else {
+        res.json({message: "match"})
+      }
+    }
+  } catch (err) {
+    console.log(err)
+  }
+}
+
+const returnStudents = async (req, res) => {
+  const userID = req.params.userID
+  const foundUser = await User.findById({_id: userID}).lean().catch(err => {       // lean(): foundUser can be modified locally (not attached to the db)
+    console.log("Finding user error when returning students: " + err)
+  })
+  if (foundUser != null) {
+      res.json({
+        message: "success",
+        students: foundUser.students
+      })
+  } else {
+    res.json({
+      message: "invalid",
+      profiles: null
+    })
+  }
+}
+
+
 const verifyEmailForReset = async (req, res) => {
+  if (!req.body.email) {
+    return res.json({message: "missing body"})
+  }
   const foundUser = await User.findOne({username: req.body.email.toLowerCase()}).catch(err => {
     console.log("Finding user error when verifying email: " + err)
   }) // check if the account exists in database
@@ -196,10 +274,15 @@ const verifyEmailForReset = async (req, res) => {
 
 
 const resetPassword = async (req, res) => {
+  if (!req.body.otp || !req.body.password) {
+    return res.json({message: "missing body"})
+  }
+
   const {
     otp,
     password
   } = req.body
+
   const foundOtp = await Otp.findOneAndDelete({otp: otp}).catch(err => {console.log("Finding Otp error when reseting password: " + err)})
   if (foundOtp == null) {
     res.json({
@@ -224,7 +307,6 @@ const resetPassword = async (req, res) => {
 
 const returnProfiles = async (req, res) => {
   const userID = req.params.userID
-  // const imagesList = [];
   const foundUser = await User.findById({_id: userID}).lean().catch(err => {       // lean(): foundUser can be modified locally (not attached to the db)
     console.log("Finding user error when returning profiles: " + err)
   })
@@ -284,6 +366,33 @@ const authenticateUser = (req, res) => {
   }
 }
 
+const sendEmailToStudent = async (req, res) => {
+
+  if (!req.body.email || !req.body.username) {
+    return res.json({message: "missing body"})
+  }
+  const {email, username} = req.body
+  const foundUser = await User.findOne({username: email.toLowerCase()}).catch(err => {console.log("Finding user error: " + err)})
+  if (foundUser !== null) {
+    const mailOptions = {
+      from: 'readpalishere@gmail.com',
+      to: `${email}`,
+      subject: 'Lesson will begin soon, teacher is waiting for you',
+      html: `<h1>Hey student</h1><p>Please paste the teacher username: <b>${username}</b> and hit the call!</p>`
+    }
+  
+    transporter.sendMail(mailOptions, function(error){
+      if (error) {
+        console.log(error);
+      } else {
+        res.json({message: "success"})
+      }
+    })
+  }else {
+    res.json({message: "invalid"})
+  }
+}
+
 
 export {
   verifyEmailForSignup,
@@ -294,5 +403,8 @@ export {
   verifyEmailForReset,
   resetPassword,
   userLogout,
-  authenticateUser
+  authenticateUser,
+  addStudent,
+  returnStudents,
+  sendEmailToStudent
 }
